Added Auto-fire and PickupWidget
Added automatic firing when the left mouse button is held down. Added an Item class for objects that can be picked up. Added a WBP that displays information about an item and attached it to the item class. Added ability to display the WBP when the user is close enough and mouses over the item. When the user mouses away, the WBP hides. Added an enum around the rarity of the item and have the WBP displaying some number of stars based on the rarity.

A note on the way that stars are displayed: I don't think this is the best way and I'm about to refactor it. I'm committing this code in case my refactor is worse.