Added weapon swapping and item states
When the user is mousing over a weapon they may swap the weapon they are holding with another weapon. The weapon they are holding will be pushed away from the character with a small velocity.
Also added item states with the intention of moving the weapons through different states as they are available for pick up, being picked up, being held, and being dropped.