Added auto-pickup of ammo
Added the ability to pick up ammo. This required a new  Ammo class subclassed from item and a new Ammo WBP + a new backing class. Because the Ammo WBP is so different from the Weapon's WBP, I pulled the instantiation of the WBPs out of the Item class and into the Weapon and Ammo classes.
Switched up the "item pickup interpolation" to pick one of several interp locations in the user's viewport so that if the user picked up many items quickly they wouldn't all go to the same place.
There are some variables leftover that could be used to limit the number of times item Pickup and Equip sounds play, but I'd already refactored those sounds to be owned by the item class. In the tutorial Stephen uses the Character class as a singleton class which can manage the sounds. I considered using a subsystem as a singleton to manage all sounds, but I chose not to because the problem of "many pickups result in many sounds" isn't actually a problem to me; I kind of like it.