Added multiple weapon types & user inventory
Put a bunch of weapon data in a DataTable. Added a Pistol-type weapon. Allowed weapons to not be able to auto-fire.
Added an inventory and the ability to pickup weapons and switch between active weapons. If the inventory is full, picking up a new weapon will drop the current active weapon.