Added TwinBlast and Grux characters
Added footstep sounds & particle effects via a new AnimNotify.
It checks for the height of the footstep and if below a certain height plays water instead. It can do some simple checking to not reload data from the DataTable if the footstep was on the same material as the previous footstep.
Added code & AnimMontages for Death and Hitreacting.
Added a new character called TwinBlast.
Added an enemy called Grux and set up a Behavior Tree for them.
Added explosive boxes which can be shot and play a particle effect.
Added ability for PC to die.